requirements.txt
asgiref==3.3.1
click==7.1.2
cycler==0.10.0
Django==3.1.4
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
joblib==0.17.0
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
nltk==3.5
numpy==1.19.4
pandas==1.1.4
patool==1.12
Pillow==8.0.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
rarfile==4.0
regex==2020.11.13
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
certifi==2020.6.20
clang==10.0.1
click==7.1.2
cycler==0.10.0
joblib==0.17.0
nltk==3.5
jupyter==1.0.0
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.10.11
scikit-learn==0.23.2
scipy==1.5.3
seaborn==0.11.0
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tqdm==4.50.2
gensim==3.8.3
